refactor(agent-server): decouple interface from agent tars interface 3
