+++ multimodal/agent-server/package.json
@@ -23,7 +23,6 @@
   },
   "devDependencies": {
     "@multimodal/agent-server-interface": "workspace:*",
-    "@multimodal/agent-interface": "workspace:*",
     "@multimodal/agent-snapshot": "workspace:*",
     "@multimodal/agio": "workspace:*",
     "cors": "^2.8.5",
